chore(*): remove all githead from package.json (#12584)
These were/are temporary lerna artifacts that shouldn't be committed. They
[should not appear in package.json](https://github.com/lerna/lerna/issues/1880#issuecomment-456160357).

The _why_ of this PR is that I was publishing a minor release, and it
led to a publish failing because it was updating gitHead and trying to
commit the change. May as well remove this so we're isolated from that going forward.

This will cause a re-release of a _bunch_ of packages upon merging, but
that's probably OK.

<!--
  Have any questions? Check out the contributing docs at https://gatsby.dev/contribute, or
  ask in this Pull Request and a Gatsby maintainer will be happy to help :)
-->